Изменение : работа с главами в .IFO - заменил int -> UINT. Построение списка глав для SeekBar - ввел доп. проверку чтобы кол-во глав в .ifo файле совпадало в кол-вом полученных от DVD Navigator, т.к. иногда такой способ получения списка глав через .ifo не корректен, и мы просто не будем показывать их на SeekBar. Косметика в названиях функций. Изменение : При отркытиии структуры BD - отображаем так же 'BD' вместо имени файла.
git-svn-id: svn://mpc-next.ru/mpc-be/trunk@2599 d191f3aa-8897-45be-b9a4-56e5fb472b67

+++ src/filters/reader/VTSReader/VTSReader.cpp
@@ -123,8 +123,8 @@ STDMETHODIMP CVTSReader::Load(LPCOLESTR pszFileName, const AM_MEDIA_TYPE* pmt)
 	}
 
 	ChapRemoveAll();
-	for (int i=0; i<m_stream.GetChaptersCount(); i++) {
-		CString chap; chap.Format(_T("Chapter %d"), i+1);
+	for (size_t i = 0; i < m_stream.GetChaptersCount(); i++) {
+		CString chap; chap.Format(_T("Chapter %d"), i + 1);
 		ChapAppend(m_stream.GetChapterOffset(i), chap);
 	}
 
@@ -292,7 +292,7 @@ BSTR CVTSStream::GetTrackName(UINT aTrackIdx)
 	return m_vob->GetTrackName(aTrackIdx);
 }
 
-int CVTSStream::GetChaptersCount()
+UINT CVTSStream::GetChaptersCount()
 {
 	return m_vob->GetChaptersCount();
 }
